+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- RS-519: check bucket name convention when renaming bucket, [PR-616](https://github.com/reductstore/reductstore/pull/616)
+
 ## [1.12.1] - 2024-10-21
 
 ### Fixed